# postman-openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Transition summarize API",
    "version": "0.1.0"
  },
  "paths": {
...
    "/audio/transcribe": {
      "post": {
        "tags": [
          "audio",
          "transcription",
          "summarization"
        ],
        "summary": "Audio Trans",
        "description": "Transcribe an uploaded audio file to text.\n\nThe response format is determined by the `Accept` request header:\n- `application/json` (default): JSON object with transcription result and metadata.\n- `text/plain`: Raw transcription text.\n- `text/srt`: SRT file download. The `Content-Disposition` filename is derived from the\n  uploaded file name (e.g. `nagranie.mp3` → `nagranie.srt`).\n  Requires `transcription_response_format=srt` in the form data.\n\nParameters:\n- uploaded_file (UploadFile): The audio file to be transcribed.\n- lang (LANG_CODE): The language of the input audio in ISO-639-1 format.\n- transcription_response_format (WHISPER_RESPONSE_FORMAT): The desired format of the transcription response (default: SRT).\n\nRaises:\n- HTTP 400: If the uploaded file is not a valid audio file or has an unsupported extension.\n- HTTP 500: If an unexpected error occurs during transcription.",
        "operationId": "audio_trans_audio_transcribe_post",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_audio_trans_audio_transcribe_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Transcription result. Content type depends on the `Accept` request header.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiProcessingResult"
                },
                "example": {
                  "result": true,
                  "transcription": "1\n00:00:01,000 --> 00:00:05,000\nHello world",
                  "format": "srt"
                }
              },
              "text/plain": {
                "schema": {
                  "type": "string"
                },
                "example": "1\n00:00:01,000 --> 00:00:05,000\nHello world"
              },
              "text/srt": {
                "schema": {
                  "type": "string",
                  "format": "binary"
                },
                "example": "1\n00:00:01,000 --> 00:00:05,000\nHello world",
                "description": "SRT file download. Returned when `Accept: text/srt` and `transcription_response_format=srt`. The filename in `Content-Disposition` is derived from the uploaded file name (e.g. `nagranie.mp3` → `nagranie.srt`)."
              }
            }
          },
          "400": {
            "description": "Invalid file type or extension.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiProcessingResult"
                },
                "example": {
                  "result": false,
                  "error": "Invalid file type. Only audio files are accepted"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error during transcription.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiProcessingResult"
                },
                "example": {
                  "result": false,
                  "error": "An error occurred while processing the audio file"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      }
    },
    "/audio/summary": {
      "post": {
        "tags": [
          "audio",
          "transcription",
          "summarization"
        ],
        "summary": "Audio Summarize",
        "description": "Transcribe and summarize an uploaded audio file.\n\nThe response format is determined by the `Accept` request header:\n- `application/json` (default): JSON object containing the generated summary.\n- `text/plain`: Raw summary text.\n\nParameters:\n- uploaded_file (UploadFile): The audio file to be summarized.\n- type (SUMMARIZATION_TYPE): The type of summary to generate (concise, tldr, detailed).\n- lang (LANG_CODE): The language of the input audio and desired summary in ISO-639-1 format.\n\nRaises:\n- HTTP 400: If the uploaded file is not a valid audio file or has an unsupported extension.\n- HTTP 500: If an unexpected error occurs during transcription or summarization.",
        "operationId": "audio_summarize_audio_summary_post",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_audio_summarize_audio_summary_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Summary result. Content type depends on the `Accept` request header.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SummaryResult"
                },
                "example": {
                  "summary": "The speaker discusses the key points of the meeting..."
                }
              },
              "text/plain": {
                "schema": {
                  "type": "string"
                },
                "example": "The speaker discusses the key points of the meeting..."
              }
            }
          },
          "400": {
            "description": "Invalid file type or extension.",
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "error": {
                      "type": "string"
                    }
                  },
                  "type": "object"
                },
                "example": {
                  "error": "Invalid file type. Only audio files are accepted"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error during transcription or summarization.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SummaryResult"
                },
                "example": {
                  "result": false,
                  "error": "An error occurred while processing the audio file"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      }
    },
    "/audio/recursive": {
      "post": {
        "tags": [
          "audio",
          "transcription",
          "summarization"
        ],
        "summary": "Audio Recursive Transcribe",
        "operationId": "audio_recursive_transcribe_audio_recursive_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RecursiveAudioTranscriptionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Batch transcription result for all audio files found in the directory.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiProcessingResult"
                },
                "example": {
                  "result": true,
                  "text": "Processed 5 audio files. Successful: 5. Failed: 0."
                }
              }
            }
          },
          "207": {
            "description": "Partial success — some files failed to transcribe.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiProcessingResult"
                },
                "example": {
                  "result": false,
                  "error": "Some files failed to transcribe",
                  "text": "Processed 5 audio files. Successful: 3. Failed: 2. Failed files: /path/a.mp3, /path/b.wav"
                }
              }
            }
          },
          "400": {
            "description": "Directory does not exist or path is not a directory.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiProcessingResult"
                },
                "example": {
                  "result": false,
                  "error": "Directory does not exist: /some/path"
                }
              }
            }
          },
          "403": {
            "description": "Endpoint not available in non-local mode.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiProcessingResult"
                },
                "example": {
                  "result": false,
                  "error": "Endpoint not available."
                }
              }
            }
          },
          "500": {
            "description": "Internal server error during recursive transcription.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiProcessingResult"
                },
                "example": {
                  "result": false,
                  "error": "An error occurred while processing the directory"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
      "ApiProcessingResult": {
        "properties": {
          "result": {
            "type": "boolean",
            "title": "Success or error"
          },
          "error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error description"
          },
          "text": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Video transcription"
          }
        },
        "type": "object",
        "required": [
          "result",
          "error",
          "text"
        ],
        "title": "ApiProcessingResult"
      },
...
      "Body_audio_summarize_audio_summary_post": {
        "properties": {
          "uploaded_file": {
            "type": "string",
            "contentMediaType": "application/octet-stream",
            "title": "Uploaded File"
          },
          "type": {
            "$ref": "#/components/schemas/SUMMARIZATION_TYPE"
          },
          "lang": {
            "$ref": "#/components/schemas/LANG_CODE"
          }
        },
        "type": "object",
        "required": [
          "uploaded_file",
          "type",
          "lang"
        ],
        "title": "Body_audio_summarize_audio_summary_post"
      },
      "Body_audio_trans_audio_transcribe_post": {
        "properties": {
          "uploaded_file": {
            "type": "string",
            "contentMediaType": "application/octet-stream",
            "title": "Uploaded File"
          },
          "lang": {
            "$ref": "#/components/schemas/LANG_CODE"
          },
          "transcription_response_format": {
            "$ref": "#/components/schemas/WHISPER_RESPONSE_FORMAT",
            "default": "srt"
          }
        },
        "type": "object",
        "required": [
          "uploaded_file",
          "lang"
        ],
        "title": "Body_audio_trans_audio_transcribe_post"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "LANG_CODE": {
        "type": "string",
        "enum": [
          "en",
          "pl"
        ],
        "title": "LANG_CODE"
      },
...
      "SummaryResult": {
        "properties": {
          "summary": {
            "type": "string",
            "title": "Summary"
          }
        },
        "type": "object",
        "required": [
          "summary"
        ],
        "title": "SummaryResult"
      },
...
    },
    "securitySchemes": {
      "OAuth2PasswordBearer": {
        "type": "oauth2",
        "flows": {
          "password": {
            "scopes": {},
            "tokenUrl": "auth/token"
          }
        }
      }
    }
  },
  "servers": [
    {
      "url": "http://127.0.0.1:8086"
    }
  ],
...
}